Separate function_table implementations by platform.
